preliminary file refactoring
